move crontab example to bottom of usage methods and add warning about manual restarting needed
This crontab method only starts the scoreboard when the Pi boots up.  If the
scoreboard crashes for whatever reason, you'll have to restart it in some other
way, or reboot/power-cycle the Pi.  Since both the Supervisor and screen options
already listed are good for doing this manual restart, put the crontab example
after them so it's easier to refer to.

Thanks to falkyre and riffnshred for the suggestions!