fix(location):
Add cache file for location if lookup fails.  If lookup is success a location.json will be created and it will be loaded if it exists.  This file will be recreated on a restart if older than 7 days old.  If the  location lookup fails completely, it will default to Winnipeg but an error message will state that.